requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
python-multipart
shap
Flask
requests
python-dotenv
pymongo
python-socketio
